fix: overwrite model error
